+++ NEWS.md
@@ -1,5 +1,6 @@
-
 # gtsummary (development version)
+
+* Updates after {gt} package renamed `cells_style()` to `cell_text()` (#78)
 
 # gtsummary 1.1.1