+++ NEWS.md
@@ -1,3 +1,4 @@
+
 # gtsummary (development version)
 
 * Updates after {gt} package renamed `cells_style()` to `cell_text()` (#78)